requirements.txt
streamlit
pandas
plotly
numpy
stockstats
binance.client
yfinance